changes made

+++ php/changepp.php
@@ -13,13 +13,16 @@ if (isset($_POST['changePicture'])) {
     $pic = $_FILES['newPP'];
     $picName = pathinfo($pic['name'], PATHINFO_FILENAME);
     $picExt = pathinfo($pic['name'], PATHINFO_EXTENSION);
+    $picSize = $pic['size'];
     $newPicName = uniqid() . "." . $picExt;
+
 
     // notifictions
     $success = "<div class='notify'><p>Display Profile updated successfully</p></div>";
     $errorDb = "<div class='notify red'><p>Could not connect to Database, please try again later.</p></div>";
     $error = "<div class='notify red'><p>There was an error, please try again later.</p></div>";
-    $error = "<div class='notify red'><p>Please upload only .jpg | .jpeg | .png images only</p></div>";
+    $errorPic = "<div class='notify red'><p>Please upload only .jpg | .jpeg | .png images only</p></div>";
+    $errorSize = "<div class='notify red'><p>Image size is bigger than 150KB<br>Please upload a smaller image</p></div>";
 
 
     $targetDir = "../assets/profilePicture/";
@@ -28,30 +31,37 @@ if (isset($_POST['changePicture'])) {
 
     $allowedTypes = ["image/jpeg", "image/png", "image/png"];
 
-    if (in_array($pic['type'], $allowedTypes)) {
-        // upload picture to server
-        if (move_uploaded_file($pic['tmp_name'], $targetDir . $newPicName)) {
+    $fileSizeLimit = 150000;
 
-            if ($file !== 'default.png') {
-                unlink($filePathName);
-                $insertIntoDb = mysqli_query($conn, "UPDATE `admin` SET `pp` = '$newPicName' WHERE `email` = '$picEmail'; ");
+    if ($picSize < $fileSizeLimit) {
+        if (in_array($pic['type'], $allowedTypes)) {
+            // upload picture to server
+            if (move_uploaded_file($pic['tmp_name'], $targetDir . $newPicName)) {
+
+                if ($file !== 'default.png') {
+                    unlink($filePathName);
+                    $insertIntoDb = mysqli_query($conn, "UPDATE `admin` SET `pp` = '$newPicName' WHERE `email` = '$picEmail'; ");
+                } else {
+                    // insert the image name into the database
+                    $insertIntoDb = mysqli_query($conn, "UPDATE `admin` SET `pp` = '$newPicName' WHERE `email` = '$picEmail'; ");
+                }
+                if ($insertIntoDb) {
+                    $_SESSION['notify'] = $success;
+                    header("Location: {$_SERVER['HTTP_REFERER']}");
+                } else {
+                    $_SESSION['notify'] = $errorDb;
+                    header("Location: {$_SERVER['HTTP_REFERER']}");
+                }
             } else {
-                // insert the image name into the database
-                $insertIntoDb = mysqli_query($conn, "UPDATE `admin` SET `pp` = '$newPicName' WHERE `email` = '$picEmail'; ");
-            }
-            if ($insertIntoDb) {
-                $_SESSION['notify'] = $success;
-                header("Location: {$_SERVER['HTTP_REFERER']}");
-            } else {
-                $_SESSION['notify'] = $errorDb;
+                $_SESSION['notify'] = $error;
                 header("Location: {$_SERVER['HTTP_REFERER']}");
             }
         } else {
-            $_SESSION['notify'] = $error;
+            $_SESSION['notify'] = $errorPic;
             header("Location: {$_SERVER['HTTP_REFERER']}");
         }
     } else {
-        $_SESSION['notify'] = $errorPic;
+        $_SESSION['notify'] = $errorSize;
         header("Location: {$_SERVER['HTTP_REFERER']}");
-    }
+    };
 }
